Fine grain loop and tier consolodation.
Produce an event loop break during enqueue so that emitting a large
ammount of data will allow other tasks to run.

Consolodate some of the tier decorators to just `tier` which inspects
the function it decorators to decided if if needs to be wrapped as
an asyncio.coroutine.

Support for appending tiers which source the last one like a pipe.